automation of points total input, speeding up operation menu
The Qualtiy of life improvements stated in the introduction are that the points for finishing position are now calculated by the program itself, so the user does not need to input the points totals for each competitor, only now having to confirm how many retirements/non-classified finishes the were from the ePrix and whether or not a team/manufacturer car retired or customer/Nelson Piquet Jr Trophy driver retired. The operation option menu on the admin program now takes number input instead of text to speed up the process of selecting an operation, which speeds it up and puts it in line with the viewer program.